requirements.txt
boto3==1.15.16
celery==4.4.7
django
django-celery-beat==2.0.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-stubs==1.5.0
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djangorestframework-stubs==1.2.0
docopt==0.6.2
flake8==3.8.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
supervisor==4.2.0
paramiko==2.7.2
psycopg2-binary==2.8.6
twilio==6.46.0
uWSGI==2.0.19.1
